split git commit cache from git commit provider
